backend/requirements.txt
flask
flask-cors
openai
duckdb
python-dotenv